change utc_date to a Float

+++ server-side-ruby/db/migrations/1_create_trackers.rb
@@ -6,7 +6,7 @@ Sequel.migration do
       Float :latitude
       Float :longitude
       Timestamp :date
-      Integer :utc_date
+      Float :utc_date
     end
   end
